+++ Revo.Domain/ReadModel/EntityReadModel.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Revo.DataAccess.Entities;
-using Revo.Domain.Core;
 
 namespace Revo.Domain.ReadModel
 {